+++ pycm/cm.py
@@ -7,7 +7,7 @@ from .handlers import __obj_assign_handler__, __obj_file_handler__, __obj_matrix
 from .handlers import __imbalancement_handler__
 from .pycm_class_func import F_calc, IBA_calc, TI_calc, NB_calc, sensitivity_index_calc
 from .overall_funcs import weighted_kappa_calc, weighted_alpha_calc, alpha2_calc, brier_score_calc, log_loss_calc
-from .pycm_distance import DistanceType, DISTANCE_MAPPER
+from .distance import DistanceType, DISTANCE_MAPPER
 from .output import *
 from .utils import *
 from .params import *